% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mcmc_convergence}
\alias{check_mcmc_convergence}
\title{Check MCMC Convergence}
\usage{
check_mcmc_convergence(object)
}
\arguments{
\item{object}{An object of class \code{posterior}, containing MCMC samples.}
}
\description{
Checks the convergence of MCMC draws using \code{rhat} and effective sample sizes (\code{ess_bulk} and \code{ess_tail}).
}
\examples{
\dontrun{
options(bdlim_rhat_thresh = 1.05)
options(bdlim_ess_thresh = 500)

# check_mcmc_convergence(summary(bdlim4_fit))
}

}
\references{
https://projecteuclid.org/journals/bayesian-analysis/volume-16/issue-2/Rank-Normalization-Folding-and-Localization--An-Improved-R\%CB\%86-for/10.1214/20-BA1221.full
}
\keyword{internal}
