% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_name}
\alias{get_model_name}
\title{Get Model Name Based on Parameter Constraints}
\usage{
get_model_name(w_free, b_free)
}
\arguments{
\item{w_free}{A logical value. If TRUE, the 'w' parameter is free; if FALSE, it's constrained.}

\item{b_free}{A logical value. If TRUE, the 'b' parameter is free; if FALSE, it's constrained.}
}
\value{
A character string representing the model name:
\itemize{
\item "bw" if both 'w' and 'b' are free
\item "b" if only 'b' is free
\item "w" if only 'w' is free
\item "n" if neither 'w' nor 'b' is free
}
}
\description{
This function determines the model name based on whether the 'w' and 'b'
parameters are free or constrained.
}
\keyword{internal}
