% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getw.R
\name{getw}
\alias{getw}
\title{Get weight function}
\usage{
getw(object, type = "normal", ...)
}
\arguments{
\item{object}{An object of class bdlim1.}

\item{type}{Type of summary. The default is 'normal' which returns a summary with the mean projected to conform with the norm 1 constraint. Others are \code{raw} which is the same as normal but without the mean projected and \code{full} that returns all posterior samples used in the analysis (burn in removed and thinned).}

\item{...}{Not used.}
}
\value{
A data frame with either the summary of the weight function (types normal and raw) or a posterior sample of the weight function (type full).
}
\description{
Get weight function
}
