% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcompare.R
\name{modelcompare}
\alias{modelcompare}
\alias{modelcompare.default}
\alias{modelcompare.bdlim4}
\alias{modelcompare.bdlim4_0.4}
\title{Model Comparison for \code{\link[=bdlim4]{bdlim4()}} Fit}
\usage{
modelcompare(object)

\method{modelcompare}{default}(object)

\method{modelcompare}{bdlim4}(object)

\method{modelcompare}{bdlim4_0.4}(object)
}
\arguments{
\item{object}{An object of class \code{bdlim4} obtained from the \code{\link[=bdlim4]{bdlim4()}} function.}
}
\value{
A vector of model probabilities.
}
\description{
Calculate and compare models probabilities fitted by \code{\link[=bdlim4]{bdlim4()}} function.
}
