% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bdlim.R
\name{print.bdlim1}
\alias{print.bdlim1}
\title{Print Results}
\usage{
\method{print}{bdlim1}(x, ...)
}
\arguments{
\item{x}{An object of class bdlim1.}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print Results
}
