% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bdlim.R
\name{print.bdlim4}
\alias{print.bdlim4}
\title{Print Results}
\usage{
\method{print}{bdlim4}(x, ...)
}
\arguments{
\item{x}{An object of class bdlim4.}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print Results
}
