% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.bdlim4.R
\name{print.summary.bdlim4}
\alias{print.summary.bdlim4}
\title{Print Summary of bdlim4}
\usage{
\method{print}{summary.bdlim4}(x, ...)
}
\arguments{
\item{x}{An object of class summary.bdlim4.}

\item{...}{Not used.}
}
\value{
Assorted model output.
}
\description{
Print Summary of bdlim4
}
