% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bdlim4.R
\name{summary.bdlim4}
\alias{summary.bdlim4}
\title{Summarize \code{\link[=bdlim4]{bdlim4()}} Object}
\usage{
\method{summary}{bdlim4}(
  object,
  model = NULL,
  exponentiate = FALSE,
  probs = c(0.025, 0.975),
  ...
)
}
\arguments{
\item{object}{An object fitted with \code{\link[=bdlim4]{bdlim4()}}.}

\item{model}{A character string specifying which model to summarize. If \code{NULL}, the best fitting model is selected based on model probability.}

\item{exponentiate}{A logical value indicating whether to exponentiate the results if the logit link is used. Default is \code{FALSE}.}

\item{probs}{A numeric vector of probabilities to compute the quantiles. Default is \code{c(0.025, 0.975)}.}

\item{...}{Additional arguments passed to \code{posterior::summarize_draws}.}
}
\value{
A \code{summary.bdlim4} object.
}
\description{
Summarizes a fitted \code{bdlim4} object.
}
\examples{
\donttest{

library(bdlim)
library(future)

# To run in parallel, change workers to > 1
plan(multisession, workers = 1)
# run BDLIM with modification by ChildSex
fit_sex <- bdlim4(
  y = sbd_bdlim$bwgaz,
  exposure = sbd_bdlim[, paste0("pm25_", 1:37)],
  covars = sbd_bdlim[
    , c("MomPriorBMI", "MomAge", "race", "Hispanic", "EstMonthConcept", "EstYearConcept")
  ],
  group = as.factor(sbd_bdlim$ChildSex),
  df = 5,
  nits = 100 # Set to 100 for example, increase in analysis
)
plan(sequential)

#summarize results
summary(fit_sex)

# obtain estimates of the distributed lag function
# these are note displayed when printed but available for use
sfit_sex <- summary(fit_sex)
head(sfit_sex$dlfun)

# can summarize with a specific model
sfit_hisp_n <- summary(fit_sex, model="n") # no modification
sfit_hisp_b <- summary(fit_sex, model="b") # subgroup-specific effects (beta)
sfit_hisp_w <- summary(fit_sex, model="w") # subgroup-specific weight function
sfit_hisp_bw <- summary(fit_sex, model="bw") # both subgroup-specific

}
}
