% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlim4.R
\name{validate_bdlim}
\alias{validate_bdlim}
\title{Validate Common Inputs of \link{bdlim4} and \link{bdlim1}}
\usage{
validate_bdlim(
  y,
  exposure,
  covars,
  group,
  id,
  df,
  nits,
  nburn,
  nthin,
  nchains,
  family = c("GAUSSIAN", "BINOMIAL")
)
}
\arguments{
\item{y}{A vector of outcomes.}

\item{exposure}{A matrix or data.frame of exposures, with one row per individual and one column per time the exposure is observed.}

\item{covars}{A matrix or data.frame of covariates. Covariates should not include the grouping factor (see 'group'). Factor covariates may be included.}

\item{group}{A vector of group memberships. This should be a factor variable corresponding to each row of 'covars' or 'group'.}

\item{id}{An optional vector of individual IDs for cases with repeated measures or other groupings where a random intercept is needed. This must be a factor variable.}

\item{df}{Degrees of freedom for the weight functions passed to \code{\link[splines:ns]{splines::ns()}}.}

\item{nits}{Number of MCMC iterations.}

\item{nburn}{Number of MCMC iterations to be discarded as burn-in. Defaults to half the total number of iterations. Used for WAIC, but also passed to summary and plot functions.}

\item{nthin}{Thinning factor for MCMC. Used for WAIC, but also passed to summary and plot functions.}

\item{nchains}{Number of parallel chains per model. Not yet implemented.}

\item{family}{Model family to use. Supported options are "gaussian" for a normal/Gaussian linear model, and "binomial" for a logistic model.}
}
\description{
Validate Common Inputs of \link{bdlim4} and \link{bdlim1}
}
