% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlim1.R
\name{bdlim1}
\alias{bdlim1}
\title{Fit the BDLIM Model with One Pattern of Modification}
\usage{
bdlim1(
  y,
  exposure,
  covars,
  group,
  id,
  w_free,
  b_free,
  df,
  nits,
  nburn,
  nthin,
  chains,
  family,
  loglik_all
)
}
\arguments{
\item{y}{A vector of outcomes. Numeric vectors of 0s and 1s for binomial family.}

\item{exposure}{A matrix or data frame of exposures, with one row per individual and one column per time the exposure is observed.}

\item{covars}{A matrix or data frame of covariates. Covariates should not include the grouping factor (see 'group'). Factor covariates may be included.}

\item{group}{A vector of group memberships. This should be a factor variable corresponding to each row of 'covars' or 'exposure'.}

\item{id}{An optional vector of individual IDs for cases with repeated measures or other groupings where a random intercept is needed. This must be a factor variable.}

\item{w_free}{Logical indicating if the weight functions are shared by all groups (FALSE) or group-specific (TRUE).}

\item{b_free}{Logical indicating if the effect sizes are shared by all groups (FALSE) or group-specific (TRUE).}

\item{df}{Degrees of freedom for the weight functions passed to \code{\link[splines:ns]{splines::ns()}}.}

\item{nits}{Number of MCMC iterations.}

\item{nburn}{Number of MCMC iterations to be discarded as burn-in. Defaults to half the total number of iterations. Used for WAIC, but also passed to summary and plot functions.}

\item{nthin}{Thinning factor for MCMC. Used for WAIC, but also passed to summary and plot functions.}

\item{chains}{Number of parallel chains per model. Not yet implemented.}

\item{family}{Model family to use. Supported options are "gaussian" for a normal/gaussian linear model, and "binomial" for a logistic model.}

\item{loglik_all}{Save full log likelihood matrix of kept chains for each pattern (i.e., \code{fit$fit_bw$ll_all_keep}). Needed for other validation methods such as \code{loo} but can be memory intensive. Default to \code{FALSE}.}
}
\value{
A \code{bdlim1} object.
}
\description{
Warning: Use \code{\link[=bdlim4]{bdlim4()}} with the \code{model} argument to fit the desired pattern instead. This function does not validate the inputs.
}
