% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlim4.R
\name{bdlim4}
\alias{bdlim4}
\title{Fit the BDLIM Model with 4 Patterns of Modification}
\usage{
bdlim4(
  y,
  exposure,
  covars,
  group,
  model = c("all", "bw", "b", "w", "n"),
  id = NULL,
  df,
  nits,
  nburn = round(nits/2),
  nthin = 1,
  chains = 1,
  family = c("gaussian", "binomial")
)
}
\arguments{
\item{y}{A vector of outcomes.}

\item{exposure}{A matrix or data.frame of exposures, with one row per individual and one column per time the exposure is observed.}

\item{covars}{A matrix or data.frame of covariates. Covariates should not include the grouping factor (see 'group'). Factor covariates may be included.}

\item{group}{A vector of group memberships. This should be a factor variable corresponding to each row of 'covars' or 'group'.}

\item{model}{A vector of model options including any of: "all", "bw", "b", "w", "n". Defaults to "all".}

\item{id}{An optional vector of individual IDs for cases with repeated measures or other groupings where a random intercept is needed. This must be a factor variable.}

\item{df}{Degrees of freedom for the weight functions passed to \code{\link[splines:ns]{splines::ns()}}.}

\item{nits}{Number of MCMC iterations.}

\item{nburn}{Number of MCMC iterations to be discarded as burn-in. Defaults to half the total number of iterations. Used for WAIC, but also passed to summary and plot functions.}

\item{nthin}{Thinning factor for MCMC. Used for WAIC, but also passed to summary and plot functions.}

\item{chains}{Number of parallel chains per model. Not yet implemented.}

\item{family}{Model family to use. Supported options are "gaussian" for a normal/Gaussian linear model, and "binomial" for a logistic model.}
}
\value{
A list of results from each pattern of modification, including model comparison metrics.
}
\description{
Fit the BDLIM Model with 4 Patterns of Modification
}
\examples{
\donttest{
library(bdlim)
library(future)

# To run in parallel, change workers to > 1
plan(multisession, workers = 1)
# run BDLIM with modification by ChildSex
fit_sex <- bdlim4(
  y = sbd_bdlim$bwgaz,
  exposure = sbd_bdlim[, paste0("pm25_", 1:37)],
  covars = sbd_bdlim[
    , c("MomPriorBMI", "MomAge", "race", "Hispanic", "EstMonthConcept", "EstYearConcept")
  ],
  group = as.factor(sbd_bdlim$ChildSex),
  df = 5,
  nits = 100 # Set to 100 for example, increase in analysis
)
plan(sequential)

# show model comparison results
fit_sex

# summarize results
sfit_sex <- summary(fit_sex)
sfit_sex

# graph the estimated distributed lag functions for each group
plot(sfit_sex)
}
}
