% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcompare.R
\name{modelcompare}
\alias{modelcompare}
\title{Model comparison for bdlim objects}
\usage{
modelcompare(object)
}
\arguments{
\item{object}{An object of class bdlim4 obtained from the bdlim4 function.}
}
\value{
A vector of model probabilities.
}
\description{
Model comparison for bdlim objects
}
