% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bdlim4.R
\name{summary.bdlim1}
\alias{summary.bdlim1}
\title{Summary for bdlim1}
\usage{
\method{summary}{bdlim1}(object, ...)
}
\arguments{
\item{object}{An object of class bdlim1.}

\item{...}{Not used.}
}
\value{
An object of class summary.bdlim2.
}
\description{
Summary for bdlim1
}
